+++ plugins/channelrx/wdsprx/wdsprx.cpp
@@ -895,7 +895,7 @@ void WDSPRx::sendChannelSettings(
     const WDSPRxSettings& settings,
     bool force) const
 {
-    qDebug("WDSPRx::sendChannelSettings: %d pipes", pipes.size());
+    qDebug("WDSPRx::sendChannelSettings: %d pipes", (int)pipes.size());
 
     for (const auto& pipe : pipes)
     {
